--- app/contact.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Contact() {
-  return <div>Contact</div>;
-}
-
-export default Contact;
