+++ app/contact/page.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import Navbar from "../navbar";
+import Footer from "../footer";
+
+function Contact() {
+  return (
+    <>
+      <Navbar />
+      <main></main>
+      <Footer />
+    </>
+  );
+}
+
+export default Contact;
